[ion] Clean LED API

+++ ion/src/shared/dummy/led.cpp
@@ -9,7 +9,7 @@ KDColor getColor() {
 
 void setColor(KDColor c) {}
 
-void setBlinking(float period, float dutyCycle) {}
+void setBlinking(uint16_t period, float dutyCycle) {}
 
 }
 }
